separate electron-main and renderer main scripts
