+++ td.vue/src/views/ThreatModel.vue
@@ -37,9 +37,16 @@
                         </h6>
                     </template>
                     <!-- "thumbnail": "./public/content/images/thumbnail.jpg", -->
-                    
-                    <h6 class="diagram-header-text diagram-edit">
-                        {{ diagram.description }}
+                    <a href="javascript:void(0)" @click="editDiagram(diagram)">
+                        <b-img-lazy
+                            class="m-auto d-block td-diagram-thumb"
+                            :src="require(`../assets/${diagram.thumbnail ? diagram.thumbnail.split('/').pop() : 'thumbnail.jpg'}`)"
+                            :alt="diagram.title" />
+                    </a>
+                    <h6 class="diagram-description-text">
+                        <a href="javascript:void(0)" @click="editDiagram(diagram)" class="diagram-edit">
+                            {{ diagram.description }}
+                        </a>
                     </h6>
                 </b-card>
             </b-col>
@@ -75,6 +82,10 @@
     white-space: pre-wrap;
 }
 .diagram-header-text a {
+    color: $black;
+}
+
+.diagram-description-text a {
     color: $black;
 }
